Renamed Comment component to CommentInput
