--- src/modules/core/components/Comment/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Comment';
