+++ src/modules/core/components/CommentInput/CommentInput.css.d.ts
