+++ src/modules/core/components/CommentInput/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CommentInput';
