+++ src/modules/dashboard/components/ActionsPage/ActionsComponents/RecoveryAction.tsx
@@ -15,7 +15,7 @@ import ActionsPageFeed, {
   ActionsPageSystemInfo,
   ActionsPageSystemMessage,
 } from '~dashboard/ActionsPageFeed';
-import Comment from '~core/Comment';
+import CommentInput from '~core/CommentInput';
 
 import {
   useLoggedInUser,
@@ -375,7 +375,7 @@ const RecoveryAction = ({
            */}
           {currentUserName && !ethereal && (
             <div ref={bottomElementRef}>
-              <Comment
+              <CommentInput
                 transactionHash={transactionHash}
                 colonyAddress={colonyAddress}
               />
